+++ src/util/netUtil.js
@@ -1,4 +1,3 @@
-import process from 'process';
 import { get } from 'lodash';
 import nodeFetch from 'node-fetch';
 import https from 'https';
@@ -6,18 +5,12 @@ import { Common } from '@k8slens/extensions';
 import queryString from 'query-string';
 import * as strings from '../strings';
 import { logger } from './logger';
+import { skipTlsVerify } from '../constants';
 
 const { Util } = Common;
 
 let httpsAgent;
-// NOTE: there seems to be a bug with Webpack in that if we use optional chaining (`?.`)
-//  to make this statement more terse, it just removes the `?` instead of (1) leaving
-//  it there like it should (and does for the rest of the code everywhere), or (2)
-//  transpiling it to what we've explicitly used here to get around this issue
-if (
-  process.env.LEX_CC_UNSAFE_ALLOW_SELFSIGNED_CERTS &&
-  process.env.LEX_CC_UNSAFE_ALLOW_SELFSIGNED_CERTS.match(/^(true|yes|1)$/)
-) {
+if (skipTlsVerify) {
   // SECURITY: get around issues with Clouds that have self-signed certificates (typically used
   //  for internal test Clouds of various kinds)
   logger.warn(